Add files via upload
Through my exploration in the past few days with the help of my companions, I have clearly felt my improvement. I know that my level is still low in the Computer Science, but through this week of study, I become more confident with my future computer studies. I understood how to explore and learn on my own, and I feel that I will be more confident  rather than anxious when encountering academic problems in the future.Thanks to this task-based interview, it lightened my way to the Computer Science. I learned about github, git, vscode, and so on .This interview gave me a quick overview of some complex tools, which was extremely meaningful for me.